add id maps to new file server
